refactor(downloader): extract SlskdDownload's orthogonal concerns into collaborators
SOLID-review follow-up (SRP): the adapter fused download orchestration
with ledger stewardship, two-phase source teardown, and staged-path
resolution in one 469-line class. Those are now composed collaborators,
built by the adapter from its own injected dependencies:

- TransferLedger (transfer-ledger.ts) — write-ahead ownership rows, GUID
  capture, confirmed-removal marking; best-effort by design.
- TransferTeardown (teardown.ts) — the bounded cancel→poll→remove rounds
  against slskd's Completed-flag-guarded remove.
- StagedFileResolver (staged-files.ts) — events-log paging that reports
  completed files at the real slskd-written path, re-rooted onto
  STAGING_ROOT.

SlskdDownload keeps what is genuinely its own: enqueue, the progress
poll loop, stall/queue-timeout policy detection, and outcome
aggregation. Shared vocabulary moved beside its concern: OwnedTransfer
to transfers.ts, the downloads-path helper to client.ts (now also used
by the sweep), and the ownership-narrowed poll to poll.ts — one copy
instead of the near-identical privates review flagged. Review polish:
StagedFileResolver.stagedFiles takes OwnedTransfer[] (dropping a
non-null assertion), the teardown doc scopes its best-effort claim to
removal faults, and the split D1/D2/D3 design references are anchored
with their change names.

Pure refactor: the constructor signature and all behavior are unchanged,
every existing test passes untouched.
